Poly-commitment/IPA: introduce PBT on the SRS trait
Moving the regression test regarding the number of expected chunks into this PBT
module.
Transforming it into a PBT helps testing the output with the different PCS
available in the library.
